+++ Projects/Feature/Home/Sources/Home/HomeView.swift
@@ -9,8 +9,6 @@ import SwiftUI
 
 import ComposableArchitecture
 import FeatureHomeInterface
-import FeatureProofPhotoInterface
-import SharedDesignSystem
 import SharedPerfTestingSupport
 
 /// 홈 화면을 렌더링하는 View입니다.
@@ -81,474 +79,5 @@ public struct HomeView: View {
         }
         .toolbar(.hidden, for: .navigationBar)
         .toolbar(.hidden, for: .tabBar)
-    }
-}
-
-// MARK: - HomeNavigationBarSection
-
-/// Reads `mainTitle`, `calendarMonthTitle`, `isRefreshHidden`,
-/// `hasUnreadNotification`. Isolated so changes to content / presentation
-/// fields do not invalidate the nav bar.
-private struct HomeNavigationBarSection: View {
-    let store: StoreOf<HomeReducer>
-
-    var body: some View {
-        TXNavigationBar(
-            style: .home(
-                .init(
-                    subTitle: store.calendarMonthTitle,
-                    mainTitle: store.mainTitle,
-                    isHiddenRefresh: store.isRefreshHidden,
-                    isRemainedAlarm: store.hasUnreadNotification
-                )
-            ), onAction: { action in
-                store.send(.navigationBarAction(action))
-            }
-        )
-    }
-}
-
-// MARK: - HomeCalendarSection
-
-/// Reads `$calendarDate` (binding) and `calendarWeeks`. Calendar-month
-/// probe marker stays inside this sub-view so the value read does NOT
-/// leak into the parent body.
-private struct HomeCalendarSection: View {
-    @Bindable var store: StoreOf<HomeReducer>
-
-    var body: some View {
-        let calendarView = TXCalendar(
-            mode: .weekly,
-            currentDate: $store.calendarDate,
-            weeks: store.calendarWeeks,
-            config: .init(
-                dateStyle: .init(lastDateTextColor: Color.Gray.gray500)
-            ),
-            onSelect: { item in
-                store.send(.calendarDateSelected(item))
-            },
-            onSwipe: { swipe in
-                store.send(.weekCalendarSwipe(swipe))
-            }
-        )
-        .frame(maxWidth: .infinity, maxHeight: 76)
-        .perfControl(slug: "home", element: "calendar")
-
-        if UITestMode.isProbeScenario {
-            calendarView.perfStateMarker(
-                slug: "home",
-                key: "calendar-month",
-                value: "\(store.calendarDate.year)-\(store.calendarDate.month)"
-            )
-        } else {
-            calendarView
-        }
-    }
-}
-
-// MARK: - HomeContentSection
-
-/// Reads `items`, `goalSectionTitle`. Owns the 50/200-cell `LazyVStack`
-/// whose ForEach is the dominant rendering cost. Presentation flag
-/// changes (toast / sheets / modal / alert) do NOT invalidate this view
-/// because they live in `HomePresentationLayer`.
-private struct HomeContentSection: View {
-    let store: StoreOf<HomeReducer>
-
-    #if PERF_TESTING
-    /// Pass 4-S2 — guards the self-run scroll Task so it only fires once per
-    /// scene appearance, even if SwiftUI invalidates the view multiple times
-    /// during initial layout settling.
-    @State private var selfRunScrollStarted: Bool = false
-    /// Pass 4-S2 — flips to `"true"` once the self-run scrollTo sequence
-    /// completes. Surfaced via `perfStateMarker` so trace analysis can isolate
-    /// the post-scroll window.
-    @State private var selfRunScrollDone: String = "false"
-    #endif
-
-    var body: some View {
-        #if PERF_TESTING
-        if UITestMode.isEnabled, UITestMode.isSwiftUISelfRunFeedScroll {
-            selfRunScrollContent
-        } else {
-            scrollContent
-        }
-        #else
-        scrollContent
-        #endif
-    }
-
-    private var scrollContent: some View {
-        ScrollView {
-            Group {
-                HomeHeaderRow(store: store)
-                cardList
-            }
-            .padding(.horizontal, 20)
-            .padding(.top, 16)
-            .padding(.bottom, 103)
-        }
-        .refreshable {
-            store.send(.fetchGoals)
-        }
-    }
-
-    #if PERF_TESTING
-    /// Pass 4-S2 — Example/perf-only branch. Wraps the same `scrollContent`
-    /// in a `ScrollViewReader` (public SwiftUI API) so a self-running Task
-    /// can call `proxy.scrollTo(id:anchor:)` across a stride of
-    /// `home-heavy` item ids. State-driven self-run: not equivalent to
-    /// real finger drag; lower bound on real scroll cost.
-    private var selfRunScrollContent: some View {
-        ScrollViewReader { proxy in
-            scrollContent
-                .perfStateMarker(
-                    slug: "home",
-                    key: "swiftui-selfrun-scroll",
-                    value: selfRunScrollDone
-                )
-                .onAppear { startSelfRunScrollIfNeeded(proxy: proxy) }
-        }
-    }
-
-    private func startSelfRunScrollIfNeeded(proxy: ScrollViewProxy) {
-        guard !selfRunScrollStarted else { return }
-        selfRunScrollStarted = true
-        let allIds = store.items.map(\.id)
-        let stridedTargets = stride(from: 5, to: allIds.count, by: 5)
-            .compactMap { allIds.indices.contains($0) ? allIds[$0] : nil }
-        let preRollNanos: UInt64 = 1_000_000_000
-        let stepIntervalNanos: UInt64 = 300_000_000
-        Task { @MainActor in
-            try? await Task.sleep(nanoseconds: preRollNanos)
-            for id in stridedTargets {
-                withAnimation(.easeInOut(duration: 0.25)) {
-                    proxy.scrollTo(id, anchor: .top)
-                }
-                try? await Task.sleep(nanoseconds: stepIntervalNanos)
-            }
-            selfRunScrollDone = "true"
-        }
-    }
-    #endif
-
-    var cardList: some View {
-        LazyVStack(spacing: 16) {
-            ForEach(store.items) { item in
-                goalCard(for: item)
-                    .perfCell(slug: "home", stableId: item.id)
-            }
-        }
-        .padding(.top, 12)
-        .perfFeed("home")
-    }
-
-    func goalCard(for item: HomeGoalItem) -> some View {
-        GoalCardView(
-            item: item.card,
-            onHeaderTapped: { store.send(.headerTapped(item.card)) },
-            onCheckButtonTapped: {
-                store.send(.goalCheckButtonTapped(
-                    id: item.id,
-                    isChecked: item.card.myCard.isSelected
-                ))
-            },
-            actionLeft: { store.send(.myCardTapped(item.card)) },
-            actionRight: { store.send(.yourCardTapped(item.card)) }
-        )
-    }
-}
-
-// MARK: - HomeHeaderRow
-
-/// Isolated so `goalSectionTitle` re-computation only invalidates this
-/// small Text row, not the entire content section or the card list.
-private struct HomeHeaderRow: View {
-    let store: StoreOf<HomeReducer>
-
-    var body: some View {
-        HStack(spacing: 0) {
-            Text(store.goalSectionTitle)
-                .typography(.b1_14b)
-
-            Spacer()
-
-            Button {
-                store.send(.editButtonTapped)
-            } label: {
-                Text("편집")
-                    .typography(.b1_14b)
-                    .foregroundStyle(Color.Gray.gray500)
-            }
-        }
-        .frame(height: 24)
-    }
-}
-
-// MARK: - HomeEmptyContentSection
-
-/// Reads `hadFirstGoal`. Manages local `emptyScrollHeight` `@State` so it
-/// doesn't get reset when content sections re-render.
-private struct HomeEmptyContentSection: View {
-    let store: StoreOf<HomeReducer>
-
-    @State private var emptyScrollHeight: CGFloat = 0
-
-    var body: some View {
-        VStack(spacing: 0) {
-            HomeHeaderRow(store: store)
-                .padding(.horizontal, 20)
-                .padding(.top, 16)
-
-            ScrollView {
-                goalEmptyView
-                    // 실제 가시 영역 기준으로 중앙 정렬되도록 탭바 높이만큼 차감
-                    .frame(maxWidth: .infinity, minHeight: max(0, emptyScrollHeight - 58))
-                    .padding(.bottom, 58)
-            }
-            .scrollIndicators(.hidden)
-            .refreshable {
-                store.send(.fetchGoals)
-            }
-            .overlay(alignment: .bottomTrailing) {
-                emptyArrow
-            }
-            .frame(maxHeight: .infinity)
-            .background {
-                GeometryReader { geo in
-                    Color.clear
-                        .onAppear { emptyScrollHeight = geo.size.height }
-                        .onChange(of: geo.size.height) { _, newValue in
-                            emptyScrollHeight = newValue
-                        }
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    var goalEmptyView: some View {
-        Group {
-            if store.hadFirstGoal == true {
-                VStack(spacing: 8) {
-                    Image.Illustration.scare
-                        .resizable()
-                        .frame(width: 164, height: 164)
-
-                    Text("이 날은 목표가 없어요!")
-                        .typography(.t2_16b)
-                        .foregroundStyle(Color.Gray.gray400)
-                }
-            } else if store.hadFirstGoal == false {
-                VStack(spacing: 0) {
-                    Image.Illustration.emptyPoke
-                        .frame(height: 116)
-
-                    Text("첫 목표를 세워볼까요?")
-                        .typography(.t2_16b)
-                        .foregroundStyle(Color.Gray.gray400)
-                        .padding(.top, 16)
-
-                    Text("+ 버튼을 눌러 목표를 추가해보세요")
-                        .typography(.c1_12r)
-                        .foregroundStyle(Color.Gray.gray300)
-                        .padding(.top, 4)
-                }
-            }
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-
-    var emptyArrow: some View {
-        Image.Illustration.arrow
-            .padding(.bottom, 71 + 58)
-            .padding(.trailing, 86)
-            .ignoresSafeArea()
-    }
-}
-
-// MARK: - HomePresentationLayer
-
-/// Owns ALL presentation modifiers (bottom sheets, modal, fullScreenCover,
-/// alert) and their bindings to `$store.isAddGoalPresented`,
-/// `$store.isCalendarSheetPresented`, `$store.calendarSheetDate`,
-/// `$store.modal`, `$store.isProofPhotoPresented`, `$store.proofPhoto`
-/// (scope), and `$store.isCameraPermissionAlertPresented`. SwiftUI's
-/// `@ObservableState` tracking scopes those reads to this modifier's body,
-/// so a presentation flag flip does not invalidate `HomeContentSection` or
-/// `HomeNavigationBarSection`.
-private struct HomePresentationLayer: ViewModifier {
-    @Bindable var store: StoreOf<HomeReducer>
-    @Dependency(\.proofPhotoFactory) var proofPhotoFactory
-
-    func body(content: Content) -> some View {
-        content
-            .txBottomSheet(
-                isPresented: $store.isAddGoalPresented,
-                showDragIndicator: true,
-                sheetContent: {
-                    AddGoalListView { category in
-                        store.send(.addGoalButtonTapped(category))
-                    }
-                }
-            )
-            .txBottomSheet(
-                isPresented: $store.isCalendarSheetPresented,
-                sheetContent: {
-                    TXCalendarBottomSheet(
-                        selectedDate: $store.calendarSheetDate,
-                        completeButtonText: "완료",
-                        onComplete: {
-                            store.send(.monthCalendarConfirmTapped)
-                        }
-                    )
-                }
-            )
-            .txModal(
-                item: $store.modal,
-                onAction: { action in
-                    if action == .confirm {
-                        store.send(.modalConfirmTapped)
-                    }
-                }
-            )
-            .transaction { transaction in
-                transaction.disablesAnimations = false
-            }
-            .fullScreenCover(
-                isPresented: $store.isProofPhotoPresented,
-                onDismiss: { store.send(.proofPhotoDismissed) },
-            ) {
-                if let proofPhotoStore = store.scope(state: \.proofPhoto, action: \.proofPhoto) {
-                    proofPhotoFactory.makeView(proofPhotoStore)
-                }
-            }
-            .cameraPermissionAlert(
-                isPresented: $store.isCameraPermissionAlertPresented,
-                onDismiss: { store.send(.cameraPermissionAlertDismissed) }
-            )
-    }
-}
-
-// MARK: - HomePerfActionHarness
-
-/// PERF-only controls used by Pass 3 **probe scenarios** (toast / calendar
-/// month toggle). Extracted into its own sub-view so its reads on
-/// `store.calendarDate` (for the month-toggle buttons) don't pollute the
-/// parent `HomeView.body` read-set even when the probe scenario is active
-/// (lower-order concern — the probe scenario is not the authoritative
-/// rendering metric, but isolation is still good hygiene).
-///
-/// Production builds never enter this branch because
-/// `UITestMode.isProbeScenario` requires the `-UITEST_PROBE_SCENARIO`
-/// launch argument. Buttons use a 44x44 Text label so `XCUIElement.tap()`
-/// can resolve a valid hit point.
-///
-/// **Known limitation**: this harness is the first child of HomeView's
-/// VStack and shifts the production layout by ~44pt when the probe
-/// scenario is active. `.overlay` placement (which would be
-/// layout-neutral) produced non-deterministic `hit point {-1, -1}` for
-/// some buttons on iOS 26.2 simulator. **The harness must NOT be mixed
-/// into authoritative rendering scenarios** — rendering scenarios launch
-/// via `-UITEST_RENDERING_SCENARIO` which keeps this harness disabled.
-private struct HomePerfActionHarness: View {
-    let store: StoreOf<HomeReducer>
-
-    var body: some View {
-        HStack(spacing: 0) {
-            Button {
-                store.send(.showToast(.warning(message: "perf-toast")))
-            } label: {
-                Text(verbatim: "T")
-                    .frame(width: 44, height: 44)
-            }
-            .accessibilityIdentifier("feature.home.perf.toast-show")
-
-            Button {
-                store.toast = nil
-            } label: {
-                Text(verbatim: "X")
-                    .frame(width: 44, height: 44)
-            }
-            .accessibilityIdentifier("feature.home.perf.toast-dismiss")
-
-            Button {
-                var next = store.calendarDate
-                next.goToNextMonth()
-                store.send(.setCalendarDate(next))
-            } label: {
-                Text(verbatim: "▶")
-                    .frame(width: 44, height: 44)
-            }
-            .accessibilityIdentifier("feature.home.perf.calendar-next")
-
-            Button {
-                var prev = store.calendarDate
-                prev.goToPreviousMonth()
-                store.send(.setCalendarDate(prev))
-            } label: {
-                Text(verbatim: "◀")
-                    .frame(width: 44, height: 44)
-            }
-            .accessibilityIdentifier("feature.home.perf.calendar-prev")
-        }
-        .opacity(0.05)
-    }
-}
-
-// MARK: - PERF Toast Presentation Harness
-
-/// PERF-only modifier used by Pass 3 **probe scenarios**. Observes
-/// `store.toast` and exposes a deterministic state-change marker. In
-/// production this modifier returns `content` unchanged so HomeView's
-/// read-set never includes `toast`.
-///
-/// **Probe context**: production HomeView does not observe `toast` (the
-/// field is displayed at the MainTab level in the production app shell).
-/// This modifier adds an artificial UITEST-only observation path so the
-/// toast probe scenario can exercise observation scoping experiments. The
-/// scenario is therefore **not representative of the user's real rendering
-/// path**, and the resulting XCTest numbers must not be cited as UI
-/// Rendering improvement evidence.
-///
-/// Avoids `.txToast(item:)` because its 3-second auto-dismiss would add
-/// non-deterministic state changes during measurement. The lightweight
-/// overlay captures the same observation cost (HomeView body re-evaluating
-/// on `toast` mutation) without auto-dismiss noise.
-private struct PerfToastPresentationHarness: ViewModifier {
-    @Binding var toast: TXToastType?
-
-    func body(content: Content) -> some View {
-        if UITestMode.isProbeScenario {
-            content
-                .overlay(alignment: .bottom) {
-                    if toast != nil {
-                        Color.clear.frame(width: 1, height: 1)
-                    }
-                }
-                .perfStateMarker(
-                    slug: "home",
-                    key: "toast",
-                    value: toast == nil ? "hidden" : "visible"
-                )
-        } else {
-            content
-        }
-    }
-}
-
-/// Wraps `perfCounterMarkers` so the counter accessibility overlays only
-/// attach in probe scenarios. Rendering / smoke launches see no marker
-/// overlays at all.
-private struct PerfHomeCounterMarkersHarness: ViewModifier {
-    func body(content: Content) -> some View {
-        if UITestMode.isProbeScenario {
-            content.perfCounterMarkers(
-                slug: "home",
-                keys: ["home.view.rebuild.proxy"]
-            )
-        } else {
-            content
-        }
     }
 }
