+++ Projects/Shared/PerfTestingSupport/UITests/Sources/XCUIApplication+Perf.swift
@@ -1,21 +1,19 @@
 import XCTest
 
 public extension XCUIApplication {
-    /// Selects which Pass 3 measurement mode the app launches in.
-    /// `nil` keeps the app in plain UITest mode without any PERF harness or
-    /// probe markers — appropriate for smoke / navigation / scroll tests
-    /// that should not see harness-induced layout shifts.
+    /// 앱을 어떤 perf scenario mode로 실행할지 나타냅니다.
+    /// `nil`이면 PERF harness나 probe marker 없이 plain UITest mode로 실행합니다.
     enum PerfScenarioKind: String {
-        /// Activates the PERF action harness, probe markers, and proxy
-        /// counters. For driver/marker sanity tests only — NOT for
-        /// authoritative rendering measurements.
+        /// PERF action harness, probe marker, proxy counter를 켭니다.
+        /// Driver / marker sanity test 전용이며 authoritative rendering 측정에는 사용하지 않습니다.
         case probe = "-UITEST_PROBE_SCENARIO"
-        /// Keeps the PERF harness disabled while still being a UITest
-        /// driver. Intended for xctrace / Instruments recording of real
-        /// rendering scenarios (e.g. home-heavy feed scroll).
+        /// UITest driver는 유지하되 PERF harness는 끕니다.
+        /// xctrace / Instruments로 실제 rendering scenario를 기록할 때 사용합니다.
         case rendering = "-UITEST_RENDERING_SCENARIO"
     }
 
+    /// perf UITest용 launch argument를 구성하고 앱을 실행합니다.
+    /// seed, scenario, animation 비활성화 여부를 한 곳에서 맞추기 위한 helper입니다.
     static func launchForPerf(
         seed: String,
         scenario: PerfScenarioKind? = nil,
